+++ NEWS.md
@@ -1,5 +1,9 @@
 # gouvdown.fonts 0.0.0.9000
 
+* Splitted Spectral fonts dependency in two different dependencies for desktop 
+fonts and small caps. Use `html_dependency_spectral()` for the desktop font and
+`html_dependency_spectral_sc()` for the small caps font.
+
 * Added Spectral font v2.003 true type files and created a 
 `html_dependency_spectral()` function.